clean up logout, tests, drop `--future` references
