+++ packages/cli/src/commands/logout/index.ts
@@ -1,13 +1,6 @@
 import { printError } from '../../util/error';
-import {
-  writeToConfigFile,
-  writeToAuthConfigFile,
-} from '../../util/config/files';
 import { parseArguments } from '../../util/get-args';
 import type Client from '../../util/client';
-import { getCommandName } from '../../util/pkg-name';
-import { isAPIError } from '../../util/errors-ts';
-import { errorToString } from '@vercel/error-utils';
 import { help } from '../help';
 import { logoutCommand } from './command';
 import { getFlagsSpecification } from '../../util/get-flags-specification';
@@ -16,8 +9,6 @@ import { LogoutTelemetryClient } from '../../util/telemetry/commands/logout';
 import { logout as future } from './future';
 
 export default async function logout(client: Client): Promise<number> {
-  const { authConfig, config } = client;
-
   let parsedArgs = null;
 
   const flagsSpecification = getFlagsSpecification(logoutCommand.options);
@@ -42,54 +33,5 @@ export default async function logout(client: Client): Promise<number> {
     return 0;
   }
 
-  if (authConfig.type === 'oauth') {
-    return await future(client);
-  }
-
-  if (!authConfig.token) {
-    output.note(
-      `Not currently logged in, so ${getCommandName('logout')} did nothing`
-    );
-    return 0;
-  }
-
-  output.spinner('Logging out…', 200);
-  let exitCode = 0;
-
-  try {
-    await client.fetch('/v3/user/tokens/current', {
-      method: 'DELETE',
-      useCurrentTeam: false,
-    });
-  } catch (err: unknown) {
-    if (isAPIError(err)) {
-      if (err.status === 403) {
-        output.debug('Token is invalid so it cannot be revoked');
-      } else if (err.status !== 200) {
-        output.debug(err?.message ?? '');
-        exitCode = 1;
-      }
-    }
-  }
-
-  delete config.currentTeam;
-
-  delete authConfig.token;
-
-  try {
-    writeToConfigFile(config);
-    writeToAuthConfigFile(authConfig);
-    output.debug('Configuration has been deleted');
-  } catch (err: unknown) {
-    output.debug(errorToString(err));
-    exitCode = 1;
-  }
-
-  if (exitCode === 0) {
-    output.log('Logged out!');
-  } else {
-    output.error('Failed during logout');
-  }
-
-  return exitCode;
+  return await future(client);
 }
